default/requirements-dev.txt
aiofiles==0.5.0
astroid==2.4.2
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.59.0
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.21.0
google-auth==1.19.0
google-cloud-core==1.3.0
google-cloud-datastore==1.13.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
h11==0.9.0
httptools==0.1.1
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-multipart==0.0.5
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
starlette==0.13.4
toml==0.10.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
